requirements.txt
torch>=2.1.0
biopython>=1.80
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
gradio>=4.0.0
